+++ src/js/core/migrator.js
@@ -352,5 +352,35 @@ const migrator = {
       await browser.storage.local.set({ configurations : configurations, schemaVersion : 5 });
       migrator.migrate();
     }
+  },
+
+  /**
+   * Version: EPG 6.5.0
+   *
+   * @returns {void}
+   */
+  async migration_5 () {
+    const { configurations } = await browser.storage.local.get({ configurations : [] });
+    const configurationLength = configurations.length;
+
+    if (configurationLength > 0) {
+      for (let i = 0; i < configurationLength; i++) {
+        const { configuration } = configurations[i];
+
+        /*
+         * @removed
+         *
+         * "UserMessaging": { "WhatsNew" }
+         */
+        if (configuration.select.UserMessaging_WhatsNew) {
+          delete configuration.select.UserMessaging_WhatsNew;
+        }
+
+        configurations[i].configuration = configuration;
+      }
+
+      await browser.storage.local.set({ configurations : configurations, schemaVersion : 6 });
+      migrator.migrate();
+    }
   }
 };