+++ CHANGELOG.md
@@ -41,8 +41,8 @@
   policy, see [#286](https://github.com/cadeyrn/enterprise-policy-generator/issues/286)
 - Starting with Firefox 139, the SearchEngines policy is no longer ESR only, see
   [#298](https://github.com/cadeyrn/enterprise-policy-generator/issues/298)
-- Removed WhatsNew option from UserMessaging policy, see
-  [#310](https://github.com/cadeyrn/enterprise-policy-generator/issues/310) (Thanks, berkaynayman!)
+- Removed WhatsNew option from UserMessaging policy (Thanks, berkaynayman!) and removed it from existing configurations,
+  see [#310](https://github.com/cadeyrn/enterprise-policy-generator/issues/310) 
 
 #### Translations